Physical register calculations.
Also changed RFREQ to be represented as an integer (long long) rather
than a double for easier physical register calcuations.  Also moved most
of the JS test into a try-finally block to keep the state of CCS the
same regardless of whether an exception is thrown.

+++ si570-register-calculator.h
@@ -10,7 +10,7 @@ class Si570RegisterCalculator {
     double get_frequency();
 
     // Returns the multiplier and divider values.
-    double get_rfreq();
+    long long get_rfreq();
     int get_n1();
     int get_hs_div();
 
@@ -38,7 +38,7 @@ class Si570RegisterCalculator {
     double current_frequency_;
 
     // Multiplier and divider values.
-    double rfreq_;
+    long long rfreq_;  // Decimal represented as a 38-bit integer.
     int n1_;
     int hs_div_;
 };
